[manuf] move FT cert operations to lib
This moves the FT certificate utility functions to a separate library to
facilitate building different provisioning test programs for different
SKUs.
